# archsmartadm.indiginfoundation.com/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchSmart API Documentation",
        "description": "Complete API documentation for ArchSmart Real Estate Platform. \r\n * \r\n * **Authentication:**\r\n * 1. First, login using the `/login` endpoint with your credentials\r\n * 2. Copy the token from the response\r\n * 3. Click the 'Authorize' button (🔓) at the top right\r\n * 4. Paste your token in the value field (just the token, without 'Bearer')\r\n * 5. Click 'Authorize' and then 'Close'\r\n * 6. Now you can test all protected endpoints\r\n * \r\n * **Note:** All API requests should use `Accept: application/json` header.",
        "contact": {
            "name": "ArchSmart Support",
            "email": "admin@archsmart.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://archsmart.indiginfoundation.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://archsmartadm.indiginfoundation.com/api",
            "description": "Production API Server"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Local Development Server"
        },
        {
            "url": "/api",
            "description": "Current Domain API Server"
        }
    ],
    "paths": {
...
        "/consultations": {
            "get": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Get all consultations (Admin only)",
                "description": "Retrieve list of consultation requests with optional limit",
                "operationId": "03192b1632d51781c7469b20e2771020",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of results (1-25)",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of consultations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "total": {
                                            "type": "integer"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Consultation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Book a consultation",
                "description": "Submit a consultation request. Public endpoint.",
                "operationId": "8814cd9ef86cdeb5714a8f93ce171ece",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "phone"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "I would like to schedule a consultation..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Consultation booked successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Consultation booked successfully!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/consultations/{consultation}": {
            "put": {
                "tags": [
                    "Consultations"
                ],
                "summary": "Respond to consultation (Admin only)",
                "description": "Update consultation status and add response message",
                "operationId": "d36c4ebe240d4ddfff002456448e7632",
                "parameters": [
                    {
                        "name": "consultation",
                        "in": "path",
                        "description": "Consultation ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "pending",
                                            "responded",
                                            "closed"
                                        ]
                                    },
                                    "response_message": {
                                        "type": "string",
                                        "example": "Thank you for your consultation request..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Consultation updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Consultation updated successfully!"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Consultation"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - Admin only"
                    },
                    "404": {
                        "description": "Consultation not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/dashboard/overview": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Get dashboard overview",
                "description": "Retrieve dashboard metrics. Admins see system-wide statistics, agents see their own property data.",
                "operationId": "651bc227644e3eaf947f107ac3d77adf",
                "responses": {
                    "200": {
                        "description": "Dashboard overview data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "totals": {
                                            "properties": {
                                                "users": {
                                                    "type": "integer"
                                                },
                                                "agents": {
                                                    "type": "integer"
                                                },
                                                "admins": {
                                                    "type": "integer"
                                                },
                                                "blog_posts": {
                                                    "type": "integer"
                                                },
                                                "properties": {
                                                    "type": "integer"
                                                },
                                                "featured_properties": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "recent_properties": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Property"
                                            }
                                        },
                                        "recent_blog_posts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BlogPost"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Property": {
                "title": "Property",
                "description": "Property listing model",
                "required": [
                    "id",
                    "title",
                    "price",
                    "status"
                ],
                "properties": {
                    "id": {
                        "description": "Property ID",
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "description": "Property title",
                        "type": "string",
                        "example": "Luxury Villa in Downtown"
                    },
                    "description": {
                        "description": "Property description",
                        "type": "string",
                        "example": "Beautiful 5 bedroom villa with pool and garden"
                    },
                    "address": {
                        "description": "Property address",
                        "type": "string",
                        "example": "123 Main St, City, Country"
                    },
                    "price": {
                        "description": "Property price",
                        "type": "number",
                        "format": "float",
                        "example": 500000
                    },
                    "bedrooms": {
                        "description": "Number of bedrooms",
                        "type": "integer",
                        "example": 5
                    },
                    "bathrooms": {
                        "description": "Number of bathrooms",
                        "type": "integer",
                        "example": 3
                    },
                    "area": {
                        "description": "Property area in square feet",
                        "type": "number",
                        "format": "float",
                        "example": 3500
                    },
                    "property_type": {
                        "description": "Type of property",
                        "type": "string",
                        "enum": [
                            "house",
                            "apartment",
                            "villa",
                            "land",
                            "commercial"
                        ],
                        "example": "villa"
                    },
                    "listing_type": {
                        "description": "Listing type",
                        "type": "string",
                        "enum": [
                            "sale",
                            "rent"
                        ],
                        "example": "sale"
                    },
                    "status": {
                        "description": "Property approval status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "approved",
                            "rejected"
                        ],
                        "example": "approved"
                    },
                    "featured": {
                        "description": "Whether property is featured",
                        "type": "boolean",
                        "example": false
                    },
                    "images": {
                        "description": "Property images",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "properties/image1.jpg",
                            "properties/image2.jpg"
                        ]
                    },
                    "agent_id": {
                        "description": "ID of the agent who listed the property",
                        "type": "integer",
                        "example": 2
                    },
                    "agent": {
                        "$ref": "#/components/schemas/User"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
            "Consultation": {
                "title": "Consultation",
                "description": "Consultation request model",
                "required": [
                    "id",
                    "name",
                    "email",
                    "phone"
                ],
                "properties": {
                    "id": {
                        "description": "Consultation ID",
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "description": "Customer name",
                        "type": "string",
                        "example": "John Smith"
                    },
                    "email": {
                        "description": "Customer email",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "phone": {
                        "description": "Customer phone number",
                        "type": "string",
                        "example": "+1234567890"
                    },
                    "service": {
                        "description": "Service requested",
                        "type": "string",
                        "example": "Smart Home Installation",
                        "nullable": true
                    },
                    "message": {
                        "description": "Consultation message",
                        "type": "string",
                        "example": "I need consultation for smart home integration",
                        "nullable": true
                    },
                    "status": {
                        "description": "Consultation status",
                        "type": "string",
                        "enum": [
                            "pending",
                            "responded",
                            "closed"
                        ],
                        "example": "pending"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-01-01T00:00:00.000000Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter your Bearer token (the token will be automatically prefixed with 'Bearer ')",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}